requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0
tqdm>=4.62.0
timm>=0.9.0
POT>=0.9.0
datasets>=2.0.0